pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hora-finance"
version = "0.1.0"
description = "Personal-first finance tooling."
readme = "README.md"
requires-python = ">=3.14,<3.15"
dependencies = [
  "flask>=3.1.0",
  "google-auth>=2.38.0",
  "google-auth-oauthlib>=1.2.1",
  "google-api-python-client>=2.170.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "ruff>=0.6.9",
]

[project.scripts]
hora-finance = "hora_finance.__main__:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py314"
line-length = 100
